PythonTextToSql/text-to-sql-chatbot/backend/requirements.txt
# FastAPI Core
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Authentication & Security
PyJWT==2.8.0
passlib[bcrypt]==1.7.4
cryptography>=3.4.0

# Configuration & Settings
pydantic>=2.7.0
pydantic-settings>=2.10.1
python-dotenv==1.0.0

# Database
sqlalchemy==2.0.23
pyodbc==5.0.1

# WebSocket & Real-time
websockets==12.0

# Utility
httpx>=0.27.0
typing-extensions>=4.12.0

# Development & Testing
pytest==7.4.3
pytest-asyncio==0.21.1

# Production Dependencies
# Optional: Add these if deploying to production
# gunicorn==21.2.0
# docker==6.1.3 